+++ sympy/core/decorators.py
@@ -4,8 +4,31 @@ SymPy core decorators.
 The purpose of this module is to expose decorators without any other
 dependencies, so that they can be easily imported anywhere in sympy/core.
 """
-
 from sympify import SympifyError, sympify
+import warnings
+
+def wraps(old_func):
+    """Copy private data from ``old_func`` to ``new_func``. """
+    def decorate(new_func):
+        new_func.__dict__.update(old_func.__dict__)
+        new_func.__module__ = old_func.__module__
+        new_func.__name__   = old_func.__name__
+        new_func.__doc__    = old_func.__doc__
+        return new_func
+    return decorate
+
+def deprecated(func):
+    """This is a decorator which can be used to mark functions
+    as deprecated. It will result in a warning being emitted
+    when the function is used."""
+    def new_func(*args, **kwargs):
+        warnings.warn("Call to deprecated function %s." % func.__name__,
+                      category=DeprecationWarning)
+        return func(*args, **kwargs)
+    new_func.__name__ = func.__name__
+    new_func.__doc__ = func.__doc__
+    new_func.__dict__.update(func.__dict__)
+    return new_func
 
 def _sympifyit(arg, retval=None):
     """decorator to smartly _sympify function arguments
