+++ sympy/utilities/__init__.py
@@ -8,7 +8,7 @@ from iterables import (iff, flatten, group, take, subsets,
 from lambdify import lambdify
 from source import source
 
-from decorator import threaded, xthreaded, deprecated, wraps
+from decorator import threaded, xthreaded
 
 from cythonutils import cythonized
 from timeutils import timed
