+++ sympy/utilities/decorator.py
@@ -1,4 +1,5 @@
-import warnings
+from sympy.core.decorators import wraps
+
 
 def threaded_factory(func, use_add):
     """A factory for ``threaded`` decorators. """
@@ -61,26 +62,3 @@ def xthreaded(func):
 
     """
     return threaded_factory(func, False)
-
-def deprecated(func):
-    """This is a decorator which can be used to mark functions
-    as deprecated. It will result in a warning being emitted
-    when the function is used."""
-    def new_func(*args, **kwargs):
-        warnings.warn("Call to deprecated function %s." % func.__name__,
-                      category=DeprecationWarning)
-        return func(*args, **kwargs)
-    new_func.__name__ = func.__name__
-    new_func.__doc__ = func.__doc__
-    new_func.__dict__.update(func.__dict__)
-    return new_func
-
-def wraps(old_func):
-    """Copy private data from ``old_func`` to ``new_func``. """
-    def decorate(new_func):
-        new_func.__dict__.update(old_func.__dict__)
-        new_func.__module__ = old_func.__module__
-        new_func.__name__   = old_func.__name__
-        new_func.__doc__    = old_func.__doc__
-        return new_func
-    return decorate
