+++ sympy/utilities/tests/test_decorator.py
@@ -1,8 +1,9 @@
-from sympy.utilities.decorator import threaded, xthreaded, wraps
+from sympy.utilities.decorator import threaded, xthreaded
 
 from sympy import symbols, Eq, Matrix
 
 from sympy.abc import x, y
+from sympy.core.decorators import wraps
 
 def test_threaded():
     @threaded
